+++ package.json
@@ -17,6 +17,7 @@
 		"google-auth-library": "^5.1.1",
 		"googleapis": "^42.0.0",
 		"http-errors": "~1.6.3",
+		"jsonwebtoken": "^8.5.1",
 		"mocha": "^6.2.0",
 		"mongoose": "^5.6.5",
 		"morgan": "~1.9.1",
